made changes to theses routes, and create helper functions so that the database calls may be made internally.
