+++ backend/routes/catalogues.js
@@ -1,91 +1,54 @@
-const { Catalogue } = require("../dist/entity");
 const express = require("express");
 const router = express.Router();
-const myDatabase = require("../config/db");
+const cataloguesHelper = require("../helperFiles/cataloguesHelper.js");
 
 // GET: Fetch all catalogues
 router.get("/", async (req, res) => {
-	try {
-		const catalogueRepository = await myDatabase.getRepository(Catalogue);
-		const catalogues = await catalogueRepository.find();
-		res.json(catalogues);
-	} catch (error) {
-		console.error("Error fetching catalogues:", error);
-		res.json({ error: error.message });
+	const catalogues = await cataloguesHelper.getAllCatalogues();
+	if (catalogues instanceof Error) {
+		res.json({ error: catalogues.message });
 	}
+	res.json(catalogues);
 });
 
 // POST: Create a new catalogue
 router.post("/", async (req, res) => {
-	const { name, description } = req.body;
-	try {
-		const catalogueRepository = await myDatabase.getRepository(Catalogue);
-		const newCatalogue = catalogueRepository.create({ name, description });
-		await catalogueRepository.save(newCatalogue);
-		res.json(newCatalogue);
-	} catch (error) {
-		console.error("Error creating new catalogue:", error);
-		res.json({ error: error.message });
+	const newCatalogue = await cataloguesHelper.createNewCatalogue(req);
+	if (newCatalogue instanceof Error) {
+		res.json({ error: newCatalogue.message });
 	}
+	res.json(newCatalogue);
 });
 
 // GET: Fetch a catalogue by ID
 router.get("/:id", async (req, res) => {
-	try {
-		const catalogueRepository = await myDatabase.getRepository(Catalogue);
-		const catalogue = await catalogueRepository.findOne({
-			where: { id: parseInt(req.params.id) },
-			relations: ["sites"],
-		});
-
-		if (catalogue) {
-			res.json(catalogue);
-		} else {
-			res.send("Catalogue not found");
-		}
-	} catch (error) {
-		console.error("Error fetching catalogue:", error);
-		res.json({ error: error.message });
+	const catalogue = await cataloguesHelper.getCatalogueFromId(req);
+	if (catalogue instanceof Error) {
+		res.json({ error: catalogue.message });
 	}
+	res.json(catalogue);
 });
 
 // PUT: Update an existing catalogue
 router.put("/:id", async (req, res) => {
-	const { id } = req.params;
-	const { name, description } = req.body;
-	try {
-		const catalogueRepository = await myDatabase.getRepository(Catalogue);
-		let catalogueToUpdate = await catalogueRepository.findOneBy({
-			id: parseInt(id),
-		});
-		if (catalogueToUpdate) {
-			catalogueToUpdate.name = name;
-			catalogueToUpdate.description = description;
-			await catalogueRepository.save(catalogueToUpdate);
-			res.json(catalogueToUpdate);
-		} else {
-			res.json({ message: "Catalogue not found" });
-		}
-	} catch (error) {
-		console.error("Error updating catalogue:", error);
-		res.json({ error: error.message });
+	const catalogueToUpdate = await cataloguesHelper.updateCatalogue(req);
+	if (catalogueToUpdate === "Catalogue not found") {
+		res.json({ message: "Catalogue not found" });
 	}
+	if (catalogueToUpdate instanceof Error) {
+		res.json({ error: catalogueToUpdate.message });
+	}
+	res.json(catalogueToUpdate);
 });
 
 // DELETE: Remove a catalogue
 router.delete("/:id", async (req, res) => {
-	const id = parseInt(req.params.id);
-	try {
-		const catalogueRepository = await myDatabase.getRepository(Catalogue);
-		const deleteResult = await catalogueRepository.delete(id);
-		if (deleteResult.affected > 0) {
-			res.send();
-		} else {
-			res.json({ message: "Catalogue not found" });
-		}
-	} catch (error) {
-		console.error("Error deleting catalogue:", error);
-		res.json({ error: error.message });
+	const result = await cataloguesHelper.deleteCatalogue(req);
+	if (result === "Catalogue not found") {
+		res.json({ message: "Catalogue not found" });
+	}
+	if (result instanceof Error) {
+		res.json({ error: result.message });
 	}
 });
